Migrate SlidePanel to RTL.
